requirements.txt
avro==1.11.0
beautifulsoup4==4.10.0
bleach==4.1.0
Django==4.0
django-allow-cidr==0.3.1
django-autocomplete-light==3.8.2
django-autoslug==1.9.8
django-bootstrap-v5==1.0.7
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-modeltranslation==0.17.3
django-q==1.3.9
django-reversion==4.0.1
django-translated-fields==0.11.3
django-webserver==1.2.0
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
drf-writable-nested==0.6.3
fastavro
lxml==4.7.1
more_itertools
natsort
ordered-set==4.0.2
psycopg2-binary==2.9.2
pydotplus==2.0.2
pyexcel==0.6.7
pyexcel-xlsx==0.6.0
pytest==6.2.5
python-dateutil==2.8.2
python-snappy
pytest-docker==0.10.3
requests==2.26.0
requests-mock==1.9.3
validators==0.18.2
waitress==2.0.0
